Added 404 page + fix editProfile stories
