--- src/components/forms/editNickname/editProfile.stories.ts
@@ -1,23 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react'
-
-import { EditNickname, FormValues } from '@/components/forms/editNickname/editProfile'
-
-const meta = {
-  component: EditNickname,
-  parameters: {
-    layout: 'centered',
-  },
-  tags: ['autodocs'],
-  title: 'Forms/EditNickname',
-} satisfies Meta<typeof EditNickname>
-
-export default meta
-type Story = StoryObj<typeof meta>
-
-export const Primary: Story = {
-  args: {
-    onFormSubmit: (data: FormValues) => {
-      console.log(data)
-    },
-  },
-}
