+++ src/pages/deckPage/deckPage.tsx
@@ -4,7 +4,7 @@ import { Header } from '@/components/layouts/header'
 import { Modal } from '@/components/ui/modal/modal'
 import { useGetCurrentUserDataQuery } from '@/services/authApi/authApi'
 
-import s from '@/pages/formPages/formPage.module.scss'
+import s from '@/pages/PageTemplate/pageTemplate.module.scss'
 
 export function DecksPage() {
   const { data } = useGetCurrentUserDataQuery()
